resolved __init__ issue and completed all tests
